+++ .storybook/main.ts
@@ -70,7 +70,9 @@ const config: StorybookConfig = {
     Object.assign(config.resolve?.alias, {
       "@jobber/components": path.resolve(__dirname, "../packages/components/src"),
       "@jobber/docx": path.resolve(__dirname, "../packages/docx/src"),
-      "@jobber/docz-tools": path.resolve(__dirname, "../packages/docz-tools/src/components")
+      "@jobber/docz-tools": path.resolve(__dirname, "../packages/docz-tools/src/components"),
+      "mdxUtils": path.resolve(__dirname, "components")
+
     });
     // Return the altered config
     return config;