+++ doczrc.js
@@ -16,9 +16,7 @@ export default {
   menu: ["Atlantis", "Patterns", "Components", "Hooks", "Design"],
   files: [
     "./docs/**/*.{md,mdx}",
-    ...packages.map(
-      pckg => `./packages/${pckg}/src/**/*.{md,mdx,!stories.mdx}`,
-    ),
+    ...packages.map(pckg => `./packages/${pckg}/src/**/*.{md,mdx}`),
     ...packages.map(pckg => `./packages/${pckg}/README.{md,mdx}`),
     ...packages.map(pckg => `./packages/${pckg}/CHANGELOG.md`),
   ],
@@ -32,6 +30,7 @@ export default {
     "./github",
     "./public",
     "**/__mocks__",
+    "**/*.stories.mdx",
     ...privateComponents,
   ],
 };