+++ packages/docz-tools/CHANGELOG.md
@@ -8,6 +8,38 @@ menu: Changelog
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.10](https://github.com/GetJobber/atlantis/compare/@jobber/docz-tools@2.0.9...@jobber/docz-tools@2.0.10) (2023-02-21)
+
+**Note:** Version bump only for package @jobber/docz-tools
+
+
+
+
+
+## [2.0.9](https://github.com/GetJobber/atlantis/compare/@jobber/docz-tools@2.0.8...@jobber/docz-tools@2.0.9) (2023-02-17)
+
+**Note:** Version bump only for package @jobber/docz-tools
+
+
+
+
+
+## [2.0.8](https://github.com/GetJobber/atlantis/compare/@jobber/docz-tools@2.0.7...@jobber/docz-tools@2.0.8) (2023-02-17)
+
+**Note:** Version bump only for package @jobber/docz-tools
+
+
+
+
+
+## [2.0.7](https://github.com/GetJobber/atlantis/compare/@jobber/docz-tools@2.0.6...@jobber/docz-tools@2.0.7) (2023-02-17)
+
+**Note:** Version bump only for package @jobber/docz-tools
+
+
+
+
+
 ## [2.0.6](https://github.com/GetJobber/atlantis/compare/@jobber/docz-tools@2.0.5...@jobber/docz-tools@2.0.6) (2023-02-16)
 
 **Note:** Version bump only for package @jobber/docz-tools